Front-end implementation

+++ BlazorAuthAPI/Controller/UsersController.cs
@@ -1,9 +1,11 @@
 ﻿using BlazorAuthAPI.Core.Data.Contexts;
 using BlazorAuthAPI.Core.Results;
+using BlazorAuthAPI.Core.User;
 using BlazorAuthAPI.Core.User.Commands;
 using BlazorAuthAPI.Core.User.Entities;
 using BlazorAuthAPI.Core.User.QueryCommand;
 using BlazorAuthAPI.Core.User.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -11,26 +13,34 @@ namespace BlazorAuthAPI.Controller
 {
     [ApiController]
     [Route("api/users")]
-    public class UsersController(IUserService userService, AppDbContext context) : ControllerBase
+    public class UsersController(IUserService userService, AppDbContext context, ICurrentUser currentUser) : ControllerBase
     {
         [HttpPost(Name = "CreateUser")]
-        public async Task<ActionResult<User>> Create([FromBody]AddNewUserCommand userRequest)
+        public async Task<ActionResult<LoginResult>> Create([FromBody] AddNewUserCommand userRequest)
         {
             var body = await userService.CreateAsync(userRequest);
-
-            return Ok(body);
+            var loginResult = new LoginResult(body);
+            return Ok(loginResult);
         }
 
+        [Authorize]
         [HttpDelete("{id:Guid}", Name = "DeleteUserById")]
         public async Task<IActionResult> Delete(Guid id)
         {
+            if (!currentUser.IsAuthenticated || !currentUser.IsAdmin)
+                return Unauthorized();
+
             await userService.DeleteAsync(id);
             return NoContent();
         }
 
+        [Authorize]
         [HttpGet(Name = "FindAllUsers")]
         public async Task<ActionResult<List<User>>> Get([FromQuery] UserQueryCommand? queryCommand)
         {
+            if (!currentUser.IsAuthenticated || !currentUser.IsAdmin)
+                return Unauthorized();
+
             queryCommand ??= new UserQueryCommand();
 
             var queryable = queryCommand.ApplyFilter(context.Users);
@@ -47,10 +57,11 @@ namespace BlazorAuthAPI.Controller
             return Ok(result);
         }
 
-        [HttpPut("{userId:guid}/change-password")]
-        public async Task<IActionResult> ChangePassword(Guid userId, [FromBody] ChangePasswordCommand command)
+        [Authorize]
+        [HttpPut("change-password")]
+        public async Task<IActionResult> ChangePassword([FromBody] ChangePasswordCommand command)
         {
-            await userService.ChangePasswordAsync(userId, command);
+            await userService.ChangePasswordAsync(currentUser.UserId, command);
 
             return NoContent();
         }
